Use auto-closeable RevWalk when creating review notes
In CreateReviewNotes.createNotes a RevWalk is created, but does not get
closed if an exception occurs when parsing the commit or marking objects
as uninteresting because the method just returns in the catch block. The
RevWalk is only closed in the later finally block that is not reached in
this case.

Enclose both blocks in a try-with-resources for the RevWalk, to ensure
that it always gets closed.

Change-Id: Ief04c620ffe664dd9df5527473fc1a5e6354aa33