Fix NPE upon commit that is not a patch set thus no change note involved
Before this fix, NullPointerException-s showed in error_log when 'ps'
was null. Creating a project with an empty commit led to that NPE, among
other likely cases.

Fix this so no attempt at creating a change note is made for a commit
that does not involve a patch set.

Log an info when such direct-push cases happen.

Change-Id: I455e0a4d58e78168ba4f5c55cd272232d4120ab3